+++ dballe/core/string-test.cc
@@ -1,4 +1,4 @@
-#include "core/tests.h"
+#include "dballe/core/tests.h"
 #include "string.h"
 
 using namespace dballe;
@@ -19,34 +19,59 @@ void Tests::register_tests() {
 add_method("url_pop_query_string", []() {
     string url = "http://example.org";
     string res;
-    res = url_pop_query_string(url, "foo");
+    wassert_false(url_pop_query_string(url, "foo", res));
     wassert(actual(res) == "");
     wassert(actual(url) == "http://example.org");
 
     url = "http://example.org?foo=bar";
-    res = url_pop_query_string(url, "foo");
+    wassert_true(url_pop_query_string(url, "foo", res));
     wassert(actual(res) == "bar");
     wassert(actual(url) == "http://example.org");
 
     url = "http://example.org?foo=bar&baz=gnu";
-    res = url_pop_query_string(url, "foo");
+    wassert_true(url_pop_query_string(url, "foo", res));
     wassert(actual(res) == "bar");
     wassert(actual(url) == "http://example.org?baz=gnu");
 
     url = "http://example.org?baz=gnu&foo=bar";
-    res = url_pop_query_string(url, "foo");
+    wassert_true(url_pop_query_string(url, "foo", res));
     wassert(actual(res) == "bar");
     wassert(actual(url) == "http://example.org?baz=gnu");
 
     url = "http://example.org?baz=gnu&foo=bar&wibble=wobble";
-    res = url_pop_query_string(url, "foo");
+    wassert_true(url_pop_query_string(url, "foo", res));
     wassert(actual(res) == "bar");
     wassert(actual(url) == "http://example.org?baz=gnu&wibble=wobble");
 
     url = "http://example.org?foo=bar&foo=baz";
-    res = url_pop_query_string(url, "foo");
+    wassert_true(url_pop_query_string(url, "foo", res));
     wassert(actual(res) == "bar");
     wassert(actual(url) == "http://example.org?foo=baz");
+
+    url = "http://example.org?foobar=bar&foo=baz";
+    wassert_true(url_pop_query_string(url, "foo", res));
+    wassert(actual(res) == "baz");
+    wassert(actual(url) == "http://example.org?foobar=bar");
+
+    url = "http://example.org?foo&bar=baz";
+    wassert_true(url_pop_query_string(url, "foo", res));
+    wassert(actual(res) == "");
+    wassert(actual(url) == "http://example.org?bar=baz");
+
+    url = "http://example.org?bar=baz&foo";
+    wassert_true(url_pop_query_string(url, "foo", res));
+    wassert(actual(res) == "");
+    wassert(actual(url) == "http://example.org?bar=baz");
+
+    url = "http://example.org?foo&";
+    wassert_true(url_pop_query_string(url, "foo", res));
+    wassert(actual(res) == "");
+    wassert(actual(url) == "http://example.org");
+
+    url = "http://example.org?foo=&";
+    wassert_true(url_pop_query_string(url, "foo", res));
+    wassert(actual(res) == "");
+    wassert(actual(url) == "http://example.org");
 });
 
 }