+++ dballe/db/db.cc
@@ -135,18 +135,6 @@ shared_ptr<DB> DB::connect_memory()
     return res;
 }
 
-std::shared_ptr<DB> DB::connect_test()
-{
-    if (default_format == Format::MEM)
-        return connect_memory();
-
-    const char* envurl = getenv("DBA_DB");
-    if (envurl != NULL)
-        return dynamic_pointer_cast<dballe::db::DB>(dballe::DB::connect_from_url(envurl));
-    else
-        return connect_from_file("test.sqlite");
-}
-
 const char* DB::default_repinfo_file()
 {
     const char* repinfo_file = getenv("DBA_REPINFO");