+++ dballe/db/tests.cc
@@ -16,24 +16,6 @@ using namespace std;
 
 namespace dballe {
 namespace tests {
-namespace {
-
-std::unique_ptr<dballe::sql::Connection> get_test_connection(const std::string& backend)
-{
-    std::string envname = "DBA_DB";
-    if (!backend.empty())
-    {
-        envname = "DBA_DB_";
-        envname += backend;
-    }
-    const char* envurl = getenv(envname.c_str());
-    if (envurl == NULL)
-        error_consistency::throwf("Environment variable %s is not set", envname.c_str());
-    return dballe::sql::Connection::create_from_url(envurl);
-}
-
-}
-
 
 impl::Messages messages_from_db(std::shared_ptr<db::Transaction> tr, const dballe::Query& query)
 {
@@ -187,13 +169,12 @@ bool BaseDBFixture<DB>::has_driver()
 template<typename DB>
 void BaseDBFixture<DB>::create_db()
 {
-    db = DB::create_db(backend);
+    db = DB::create_db(backend, true);
     /*
     if (auto d = dynamic_cast<db::v7::DB*>(db.get()))
         if (auto c = dynamic_cast<sql::SQLiteConnection*>(d->conn))
             c->trace();
     */
-    db->reset();
 }
 
 template<typename DB>
@@ -245,10 +226,11 @@ void DBFixture<DB>::populate_database(TestDataSet& data_set)
     wassert(data_set.populate_db(*this->db));
 }
 
-std::shared_ptr<dballe::db::v7::DB> V7DB::create_db(const std::string& backend)
+std::shared_ptr<dballe::db::v7::DB> V7DB::create_db(const std::string& backend, bool wipe)
 {
-    auto conn = get_test_connection(backend);
-    return std::make_shared<dballe::db::v7::DB>(move(conn));
+    auto options = DBConnectOptions::test_create(backend.c_str());
+    options->wipe = wipe;
+    return std::dynamic_pointer_cast<dballe::db::v7::DB>(dballe::DB::connect(*options));
 }