+++ dballe/sql/sql.cc
@@ -1,5 +1,6 @@
 #include "sql.h"
 #include "dballe/types.h"
+#include "dballe/db.h"
 #include "querybuf.h"
 #include "sqlite.h"
 #include "config.h"
@@ -50,13 +51,9 @@ void Connection::add_datetime(Querybuf& qb, const Datetime& dt) const
             dt.hour, dt.minute, dt.second);
 }
 
-std::unique_ptr<Connection> Connection::create_from_url(const std::string& url)
+std::unique_ptr<Connection> Connection::create(const DBConnectOptions& options)
 {
-    return create_from_url(url.c_str());
-}
-
-std::unique_ptr<Connection> Connection::create_from_url(const char* url)
-{
+    const char* url = options.url.c_str();
     if (strncmp(url, "sqlite://", 9) == 0)
     {
         unique_ptr<SQLiteConnection> conn(new SQLiteConnection);
@@ -92,14 +89,6 @@ std::unique_ptr<Connection> Connection::create_from_url(const char* url)
 #else
         throw error_unimplemented("MySQL support is not available");
 #endif
-    }
-    if (strncmp(url, "test:", 5) == 0)
-    {
-        const char* envurl = getenv("DBA_DB");
-        if (envurl != NULL)
-            return create_from_url(envurl);
-        else
-            return create_from_url("sqlite://test.sqlite");
     }
     error_consistency::throwf("unsupported url \"%s\"", url);
 }