+++ fortran/dballef.f90
@@ -44,7 +44,16 @@ CHARACTER(kind=c_char,len=1), PARAMETER :: DBA_MVC = CHAR(0)
 
 ! interface to generic dba functions
 INTERFACE
-  FUNCTION idba_presentati_orig(dbahandle, url) BIND(C,name='idba_presentati')
+  FUNCTION idba_connect_orig(dbahandle, url) BIND(C,name='idba_connect')
+  IMPORT
+  INTEGER(kind=c_int) :: dbahandle
+  CHARACTER(kind=c_char) :: url(*)
+  INTEGER(kind=c_int) :: idba_connect_orig
+  END FUNCTION idba_connect_orig
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_presentati_orig(dbahandle, url) BIND(C,name='idba_connect')
   IMPORT
   INTEGER(kind=c_int) :: dbahandle
   CHARACTER(kind=c_char) :: url(*)
@@ -53,14 +62,33 @@ INTERFACE
 END INTERFACE
 
 INTERFACE
-  FUNCTION idba_arrivederci(dbahandle) BIND(C,name='idba_arrivederci')
+  FUNCTION idba_disconnect(dbahandle) BIND(C,name='idba_disconnect')
+  IMPORT
+  INTEGER(kind=c_int) :: dbahandle
+  END FUNCTION idba_disconnect
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_arrivederci(dbahandle) BIND(C,name='idba_disconnect')
   IMPORT
   INTEGER(kind=c_int) :: dbahandle
   END FUNCTION idba_arrivederci
 END INTERFACE
 
 INTERFACE
-  FUNCTION idba_preparati_orig(dbahandle, handle, anaflag, dataflag, attrflag) BIND(C,name='idba_preparati')
+  FUNCTION idba_begin_orig(dbahandle, handle, anaflag, dataflag, attrflag) BIND(C,name='idba_begin')
+  IMPORT
+  INTEGER(kind=c_int),VALUE :: dbahandle
+  INTEGER(kind=c_int) :: handle
+  CHARACTER(kind=c_char) :: anaflag(*)
+  CHARACTER(kind=c_char) :: dataflag(*)
+  CHARACTER(kind=c_char) :: attrflag(*)
+  INTEGER(kind=c_int) :: idba_begin_orig
+  END FUNCTION idba_begin_orig
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_preparati_orig(dbahandle, handle, anaflag, dataflag, attrflag) BIND(C,name='idba_begin')
   IMPORT
   INTEGER(kind=c_int),VALUE :: dbahandle
   INTEGER(kind=c_int) :: handle
@@ -72,7 +100,18 @@ INTERFACE
 END INTERFACE
 
 INTERFACE
-  FUNCTION idba_messaggi_orig(handle, filename, mode, typ) BIND(C,name='idba_messaggi')
+  FUNCTION idba_begin_messages_orig(handle, filename, mode, typ) BIND(C,name='idba_begin_messages')
+  IMPORT
+  INTEGER(kind=c_int) :: handle
+  CHARACTER(kind=c_char) :: filename(*)
+  CHARACTER(kind=c_char) :: mode(*)
+  CHARACTER(kind=c_char) :: typ(*)
+  INTEGER(kind=c_int) :: idba_begin_messages_orig
+  END FUNCTION idba_begin_messages_orig
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_messaggi_orig(handle, filename, mode, typ) BIND(C,name='idba_begin_messages')
   IMPORT
   INTEGER(kind=c_int) :: handle
   CHARACTER(kind=c_char) :: filename(*)
@@ -83,7 +122,15 @@ INTERFACE
 END INTERFACE
 
 INTERFACE
-  FUNCTION idba_fatto(dbahandle) BIND(C,name='idba_fatto')
+  FUNCTION idba_commit(dbahandle) BIND(C,name='idba_commit')
+  IMPORT
+  INTEGER(kind=c_int) :: dbahandle
+  INTEGER(kind=c_int) :: idba_commit
+  END FUNCTION idba_commit
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_fatto(dbahandle) BIND(C,name='idba_commit')
   IMPORT
   INTEGER(kind=c_int) :: dbahandle
   INTEGER(kind=c_int) :: idba_fatto
@@ -233,7 +280,15 @@ INTERFACE
 END INTERFACE
 
 INTERFACE
-  FUNCTION idba_setcontextana(handle) BIND(C,name='idba_setcontextana')
+  FUNCTION idba_set_station_context(handle) BIND(C,name='idba_set_station_context')
+  IMPORT
+  INTEGER(kind=c_int),VALUE :: handle
+  INTEGER(kind=c_int) :: idba_set_station_context
+  END FUNCTION idba_set_station_context
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_setcontextana(handle) BIND(C,name='idba_set_station_context')
   IMPORT
   INTEGER(kind=c_int),VALUE :: handle
   INTEGER(kind=c_int) :: idba_setcontextana
@@ -345,7 +400,16 @@ END INTERFACE
 
 ! interface to action functions
 INTERFACE
-  FUNCTION idba_scopa_orig(handle, repinfofile) BIND(C,name='idba_scopa')
+  FUNCTION idba_reinit_db_orig(handle, repinfofile) BIND(C,name='idba_reinit_db')
+  IMPORT
+  INTEGER(kind=c_int),VALUE :: handle
+  CHARACTER(kind=c_char) :: repinfofile(*)
+  INTEGER(kind=c_int) :: idba_reinit_db_orig
+  END FUNCTION idba_reinit_db_orig
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_scopa_orig(handle, repinfofile) BIND(C,name='idba_reinit_db')
   IMPORT
   INTEGER(kind=c_int),VALUE :: handle
   CHARACTER(kind=c_char) :: repinfofile(*)
@@ -354,7 +418,16 @@ INTERFACE
 END INTERFACE
 
 INTERFACE
-  FUNCTION idba_quantesono(handle, count) BIND(C,name='idba_quantesono')
+  FUNCTION idba_query_stations(handle, count) BIND(C,name='idba_query_stations')
+  IMPORT
+  INTEGER(kind=c_int),VALUE :: handle
+  INTEGER(kind=c_int),INTENT(out) :: count
+  INTEGER(kind=c_int) :: idba_query_stations
+  END FUNCTION idba_query_stations
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_quantesono(handle, count) BIND(C,name='idba_query_stations')
   IMPORT
   INTEGER(kind=c_int),VALUE :: handle
   INTEGER(kind=c_int),INTENT(out) :: count
@@ -363,7 +436,15 @@ INTERFACE
 END INTERFACE
 
 INTERFACE
-  FUNCTION idba_elencamele(handle) BIND(C,name='idba_elencamele')
+  FUNCTION idba_next_station(handle) BIND(C,name='idba_next_station')
+  IMPORT
+  INTEGER(kind=c_int),VALUE :: handle
+  INTEGER(kind=c_int) :: idba_next_station
+  END FUNCTION idba_next_station
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_elencamele(handle) BIND(C,name='idba_next_station')
   IMPORT
   INTEGER(kind=c_int),VALUE :: handle
   INTEGER(kind=c_int) :: idba_elencamele
@@ -371,7 +452,16 @@ INTERFACE
 END INTERFACE
 
 INTERFACE
-  FUNCTION idba_voglioquesto(handle, count) BIND(C,name='idba_voglioquesto')
+  FUNCTION idba_query_data(handle, count) BIND(C,name='idba_query_data')
+  IMPORT
+  INTEGER(kind=c_int),VALUE :: handle
+  INTEGER(kind=c_int),INTENT(out) :: count
+  INTEGER(kind=c_int) :: idba_query_data
+  END FUNCTION idba_query_data
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_voglioquesto(handle, count) BIND(C,name='idba_query_data')
   IMPORT
   INTEGER(kind=c_int),VALUE :: handle
   INTEGER(kind=c_int),INTENT(out) :: count
@@ -380,7 +470,17 @@ INTERFACE
 END INTERFACE
 
 INTERFACE
-  FUNCTION idba_dammelo_orig(handle, param, param_len) BIND(C,name='idba_dammelo')
+  FUNCTION idba_next_data_orig(handle, param, param_len) BIND(C,name='idba_next_data')
+  IMPORT
+  INTEGER(kind=c_int),VALUE :: handle
+  CHARACTER(kind=c_char) :: param(*)
+  INTEGER(kind=c_int),VALUE :: param_len
+  INTEGER(kind=c_int) :: idba_next_data_orig
+  END FUNCTION idba_next_data_orig
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_dammelo_orig(handle, param, param_len) BIND(C,name='idba_next_data')
   IMPORT
   INTEGER(kind=c_int),VALUE :: handle
   CHARACTER(kind=c_char) :: param(*)
@@ -390,7 +490,15 @@ INTERFACE
 END INTERFACE
 
 INTERFACE
-  FUNCTION idba_prendilo(handle) BIND(C,name='idba_prendilo')
+  FUNCTION idba_insert_data(handle) BIND(C,name='idba_insert_data')
+  IMPORT
+  INTEGER(kind=c_int),VALUE :: handle
+  INTEGER(kind=c_int) :: idba_insert_data
+  END FUNCTION idba_insert_data
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_prendilo(handle) BIND(C,name='idba_insert_data')
   IMPORT
   INTEGER(kind=c_int),VALUE :: handle
   INTEGER(kind=c_int) :: idba_prendilo
@@ -398,7 +506,15 @@ INTERFACE
 END INTERFACE
 
 INTERFACE
-  FUNCTION idba_dimenticami(handle) BIND(C,name='idba_dimenticami')
+  FUNCTION idba_remove_data(handle) BIND(C,name='idba_remove_data')
+  IMPORT
+  INTEGER(kind=c_int),VALUE :: handle
+  INTEGER(kind=c_int) :: idba_remove_data
+  END FUNCTION idba_remove_data
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_dimenticami(handle) BIND(C,name='idba_remove_data')
   IMPORT
   INTEGER(kind=c_int),VALUE :: handle
   INTEGER(kind=c_int) :: idba_dimenticami
@@ -406,7 +522,16 @@ INTERFACE
 END INTERFACE
 
 INTERFACE
-  FUNCTION idba_voglioancora(handle, count) BIND(C,name='idba_voglioancora')
+  FUNCTION idba_query_attributes(handle, count) BIND(C,name='idba_query_attributes')
+  IMPORT
+  INTEGER(kind=c_int),VALUE :: handle
+  INTEGER(kind=c_int),INTENT(out) :: count
+  INTEGER(kind=c_int) :: idba_query_attributes
+  END FUNCTION idba_query_attributes
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_voglioancora(handle, count) BIND(C,name='idba_query_attributes')
   IMPORT
   INTEGER(kind=c_int),VALUE :: handle
   INTEGER(kind=c_int),INTENT(out) :: count
@@ -415,7 +540,17 @@ INTERFACE
 END INTERFACE
 
 INTERFACE
-  FUNCTION idba_ancora_orig(handle, param, param_len) BIND(C,name='idba_ancora')
+  FUNCTION idba_next_attribute_orig(handle, param, param_len) BIND(C,name='idba_next_attribute')
+  IMPORT
+  INTEGER(kind=c_int),VALUE :: handle
+  CHARACTER(kind=c_char) :: param(*)
+  INTEGER(kind=c_int),VALUE :: param_len
+  INTEGER(kind=c_int) :: idba_next_attribute_orig
+  END FUNCTION idba_next_attribute_orig
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_ancora_orig(handle, param, param_len) BIND(C,name='idba_next_attribute')
   IMPORT
   INTEGER(kind=c_int),VALUE :: handle
   CHARACTER(kind=c_char) :: param(*)
@@ -425,7 +560,15 @@ INTERFACE
 END INTERFACE
 
 INTERFACE
-  FUNCTION idba_critica(handle) BIND(C,name='idba_critica')
+  FUNCTION idba_insert_attributes(handle) BIND(C,name='idba_insert_attributes')
+  IMPORT
+  INTEGER(kind=c_int),VALUE :: handle
+  INTEGER(kind=c_int) :: idba_insert_attributes
+  END FUNCTION idba_insert_attributes
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_critica(handle) BIND(C,name='idba_insert_attributes')
   IMPORT
   INTEGER(kind=c_int),VALUE :: handle
   INTEGER(kind=c_int) :: idba_critica
@@ -433,7 +576,15 @@ INTERFACE
 END INTERFACE
 
 INTERFACE
-  FUNCTION idba_scusa(handle) BIND(C,name='idba_scusa')
+  FUNCTION idba_remove_attributes(handle) BIND(C,name='idba_remove_attributes')
+  IMPORT
+  INTEGER(kind=c_int),VALUE :: handle
+  INTEGER(kind=c_int) :: idba_remove_attributes
+  END FUNCTION idba_remove_attributes
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_scusa(handle) BIND(C,name='idba_remove_attributes')
   IMPORT
   INTEGER(kind=c_int),VALUE :: handle
   INTEGER(kind=c_int) :: idba_scusa
@@ -484,7 +635,21 @@ END INTERFACE
 
 ! interface to pretty-printing functions
 INTERFACE
-  FUNCTION idba_spiegal_orig(handle, ltype1, l1, ltype2, l2, res, res_len) BIND(C,name='idba_spiegal')
+  FUNCTION idba_describe_level_orig(handle, ltype1, l1, ltype2, l2, res, res_len) BIND(C,name='idba_describe_level')
+  IMPORT
+  INTEGER(kind=c_int),VALUE :: handle
+  INTEGER(kind=c_int),VALUE :: ltype1
+  INTEGER(kind=c_int),VALUE :: l1
+  INTEGER(kind=c_int),VALUE :: ltype2
+  INTEGER(kind=c_int),VALUE :: l2
+  CHARACTER(kind=c_char) :: res(*)
+  INTEGER(kind=c_int),VALUE :: res_len
+  INTEGER(kind=c_int) :: idba_describe_level_orig
+  END FUNCTION idba_describe_level_orig
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_spiegal_orig(handle, ltype1, l1, ltype2, l2, res, res_len) BIND(C,name='idba_describe_level')
   IMPORT
   INTEGER(kind=c_int),VALUE :: handle
   INTEGER(kind=c_int),VALUE :: ltype1
@@ -498,7 +663,20 @@ INTERFACE
 END INTERFACE
 
 INTERFACE
-  FUNCTION idba_spiegat_orig(handle, ptype, p1, p2, res, res_len) BIND(C,name='idba_spiegat')
+  FUNCTION idba_describe_timerange_orig(handle, ptype, p1, p2, res, res_len) BIND(C,name='idba_describe_timerange')
+  IMPORT
+  INTEGER(kind=c_int),VALUE :: handle
+  INTEGER(kind=c_int),VALUE :: ptype
+  INTEGER(kind=c_int),VALUE :: p1
+  INTEGER(kind=c_int),VALUE :: p2
+  CHARACTER(kind=c_char) :: res(*)
+  INTEGER(kind=c_int),VALUE :: res_len
+  INTEGER(kind=c_int) :: idba_describe_timerange_orig
+  END FUNCTION idba_describe_timerange_orig
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_spiegat_orig(handle, ptype, p1, p2, res, res_len) BIND(C,name='idba_describe_timerange')
   IMPORT
   INTEGER(kind=c_int),VALUE :: handle
   INTEGER(kind=c_int),VALUE :: ptype
@@ -511,7 +689,19 @@ INTERFACE
 END INTERFACE
 
 INTERFACE
-  FUNCTION idba_spiegab_orig(handle, varcode, var, res, res_len) BIND(C,name='idba_spiegab')
+  FUNCTION idba_describe_var_orig(handle, varcode, var, res, res_len) BIND(C,name='idba_describe_var')
+  IMPORT
+  INTEGER(kind=c_int),VALUE :: handle
+  CHARACTER(kind=c_char) :: varcode(*)
+  CHARACTER(kind=c_char) :: var(*)
+  CHARACTER(kind=c_char) :: res(*)
+  INTEGER(kind=c_int),VALUE :: res_len
+  INTEGER(kind=c_int) :: idba_describe_var_orig
+  END FUNCTION idba_describe_var_orig
+END INTERFACE
+
+INTERFACE
+  FUNCTION idba_spiegab_orig(handle, varcode, var, res, res_len) BIND(C,name='idba_describe_var')
   IMPORT
   INTEGER(kind=c_int),VALUE :: handle
   CHARACTER(kind=c_char) :: varcode(*)
@@ -520,14 +710,6 @@ INTERFACE
   INTEGER(kind=c_int),VALUE :: res_len
   INTEGER(kind=c_int) :: idba_spiegab_orig
   END FUNCTION idba_spiegab_orig
-END INTERFACE
-
-INTERFACE
-  FUNCTION idba_test_input_to_output(dbahandle) BIND(C,name='idba_test_input_to_output')
-  IMPORT
-  INTEGER(kind=c_int),VALUE :: dbahandle
-  INTEGER(kind=c_int) :: idba_test_input_to_output
-  END FUNCTION idba_test_input_to_output
 END INTERFACE
 
 
@@ -621,6 +803,28 @@ idba_presentati = idba_presentati_orig(dbahandle, fchartrimtostr(url))
 
 END FUNCTION idba_presentati
 
+FUNCTION idba_connect(dbahandle, url)
+INTEGER(kind=c_int) :: dbahandle
+CHARACTER(kind=c_char,len=*) :: url
+INTEGER(kind=c_int) :: idba_connect
+
+idba_connect = idba_connect_orig(dbahandle, fchartrimtostr(url))
+
+END FUNCTION idba_connect
+
+FUNCTION idba_begin(dbahandle, handle, anaflag, dataflag, attrflag)
+INTEGER(kind=c_int) :: dbahandle
+INTEGER(kind=c_int) :: handle
+CHARACTER(kind=c_char,len=*) :: anaflag
+CHARACTER(kind=c_char,len=*) :: dataflag
+CHARACTER(kind=c_char,len=*) :: attrflag
+INTEGER(kind=c_int) :: idba_begin
+
+idba_begin = idba_begin_orig(dbahandle, handle, fchartrimtostr(anaflag), &
+ fchartrimtostr(dataflag), fchartrimtostr(attrflag))
+
+END FUNCTION idba_begin
+
 FUNCTION idba_preparati(dbahandle, handle, anaflag, dataflag, attrflag)
 INTEGER(kind=c_int) :: dbahandle
 INTEGER(kind=c_int) :: handle
@@ -633,6 +837,18 @@ idba_preparati = idba_preparati_orig(dbahandle, handle, fchartrimtostr(anaflag),
  fchartrimtostr(dataflag), fchartrimtostr(attrflag))
 
 END FUNCTION idba_preparati
+
+FUNCTION idba_begin_messages(handle, filename, mode, typ)
+INTEGER(kind=c_int) :: handle
+CHARACTER(kind=c_char,len=*) :: filename
+CHARACTER(kind=c_char,len=*) :: mode
+CHARACTER(kind=c_char,len=*) :: typ
+INTEGER(kind=c_int) :: idba_begin_messages
+
+idba_begin_messages = idba_begin_messages_orig(handle, fchartrimtostr(filename), &
+ fchartrimtostr(mode), fchartrimtostr(typ))
+
+END FUNCTION idba_begin_messages
 
 FUNCTION idba_messaggi(handle, filename, mode, typ)
 INTEGER(kind=c_int) :: handle
@@ -759,6 +975,15 @@ END FUNCTION idba_unset
 
 
 ! fortran-style interface to action functions
+FUNCTION idba_reinit_db(handle, repinfofile)
+INTEGER(kind=c_int) :: handle
+CHARACTER(kind=c_char,len=*) :: repinfofile
+INTEGER(kind=c_int) :: idba_reinit_db
+
+idba_reinit_db = idba_reinit_db_orig(handle, fchartrimtostr(repinfofile))
+
+END FUNCTION idba_reinit_db
+
 FUNCTION idba_scopa(handle, repinfofile)
 INTEGER(kind=c_int) :: handle
 CHARACTER(kind=c_char,len=*) :: repinfofile
@@ -768,6 +993,15 @@ idba_scopa = idba_scopa_orig(handle, fchartrimtostr(repinfofile))
 
 END FUNCTION idba_scopa
 
+FUNCTION idba_next_data(handle, param)
+INTEGER(kind=c_int) :: handle
+CHARACTER(kind=c_char,len=*) :: param
+INTEGER(kind=c_int) :: idba_next_data
+
+idba_next_data = idba_next_data_orig(handle, param, LEN(param))
+
+END FUNCTION idba_next_data
+
 FUNCTION idba_dammelo(handle, param)
 INTEGER(kind=c_int) :: handle
 CHARACTER(kind=c_char,len=*) :: param
@@ -776,6 +1010,15 @@ INTEGER(kind=c_int) :: idba_dammelo
 idba_dammelo = idba_dammelo_orig(handle, param, LEN(param))
 
 END FUNCTION idba_dammelo
+
+FUNCTION idba_next_attribute(handle, param)
+INTEGER(kind=c_int) :: handle
+CHARACTER(kind=c_char,len=*) :: param
+INTEGER(kind=c_int) :: idba_next_attribute
+
+idba_next_attribute = idba_next_attribute_orig(handle, param, LEN(param))
+
+END FUNCTION idba_next_attribute
 
 FUNCTION idba_ancora(handle, param)
 INTEGER(kind=c_int) :: handle
@@ -839,6 +1082,19 @@ END FUNCTION idba_messages_write_next
 
 
 ! fortran-style interface to pretty-printing functions
+FUNCTION idba_describe_level(handle, ltype1, l1, ltype2, l2, res)
+INTEGER(kind=c_int) :: handle
+INTEGER(kind=c_int) :: ltype1
+INTEGER(kind=c_int) :: l1
+INTEGER(kind=c_int) :: ltype2
+INTEGER(kind=c_int) :: l2
+CHARACTER(kind=c_char,len=*),INTENT(out) :: res
+INTEGER(kind=c_int) :: idba_describe_level
+
+idba_describe_level = idba_describe_level_orig(handle, ltype1, l1, ltype2, l2, res, LEN(res))
+
+END FUNCTION idba_describe_level
+
 FUNCTION idba_spiegal(handle, ltype1, l1, ltype2, l2, res)
 INTEGER(kind=c_int) :: handle
 INTEGER(kind=c_int) :: ltype1
@@ -852,6 +1108,18 @@ idba_spiegal = idba_spiegal_orig(handle, ltype1, l1, ltype2, l2, res, LEN(res))
 
 END FUNCTION idba_spiegal
 
+FUNCTION idba_describe_timerange(handle, ptype, p1, p2, res)
+INTEGER(kind=c_int) :: handle
+INTEGER(kind=c_int) :: ptype
+INTEGER(kind=c_int) :: p1
+INTEGER(kind=c_int) :: p2
+CHARACTER(kind=c_char,len=*),INTENT(out) :: res
+INTEGER(kind=c_int) :: idba_describe_timerange
+
+idba_describe_timerange = idba_describe_timerange_orig(handle, ptype, p1, p2, res, LEN(res))
+
+END FUNCTION idba_describe_timerange
+
 FUNCTION idba_spiegat(handle, ptype, p1, p2, res)
 INTEGER(kind=c_int) :: handle
 INTEGER(kind=c_int) :: ptype
@@ -863,6 +1131,18 @@ INTEGER(kind=c_int) :: idba_spiegat
 idba_spiegat = idba_spiegat_orig(handle, ptype, p1, p2, res, LEN(res))
 
 END FUNCTION idba_spiegat
+
+FUNCTION idba_describe_var(handle, varcode, var, res)
+INTEGER(kind=c_int) :: handle
+CHARACTER(kind=c_char,len=*),INTENT(in) :: varcode
+CHARACTER(kind=c_char,len=*),INTENT(in) :: var
+CHARACTER(kind=c_char,len=*),INTENT(out) :: res
+INTEGER(kind=c_int) :: idba_describe_var
+
+idba_describe_var = idba_describe_var_orig(handle, fchartrimtostr(varcode), &
+ fchartrimtostr(var), res, LEN(res))
+
+END FUNCTION idba_describe_var
 
 FUNCTION idba_spiegab(handle, varcode, var, res)
 INTEGER(kind=c_int) :: handle